+++ test/lightning/usage_tracking_test.exs
@@ -627,7 +627,8 @@ defmodule Lightning.UsageTrackingTest do
       assert %{
                submitted: false,
                submitted_at: nil,
-               report_date: ^expected_report_date
+               report_date: ^expected_report_date,
+               submission_status: :pending
              } = report
 
       assert %{
@@ -652,7 +653,8 @@ defmodule Lightning.UsageTrackingTest do
                id: ^inserted_id,
                submitted: false,
                submitted_at: nil,
-               report_date: ^date
+               report_date: ^date,
+               submission_status: :pending
              } = report
 
       assert %{
@@ -709,7 +711,8 @@ defmodule Lightning.UsageTrackingTest do
       assert %{
                submitted: false,
                submitted_at: nil,
-               report_date: ^expected_report_date
+               report_date: ^expected_report_date,
+               submission_status: :pending
              } = report
 
       assert %{
@@ -734,7 +737,8 @@ defmodule Lightning.UsageTrackingTest do
                id: ^inserted_id,
                submitted: false,
                submitted_at: nil,
-               report_date: ^date
+               report_date: ^date,
+               submission_status: :pending
              } = report
 
       assert %{
@@ -747,7 +751,12 @@ defmodule Lightning.UsageTrackingTest do
   describe ".update_report_submission - successful" do
     setup do
       report =
-        insert(:usage_tracking_report, submitted: false, submitted_at: nil)
+        insert(
+          :usage_tracking_report,
+          submitted: false,
+          submitted_at: nil,
+          submission_status: :pending
+        )
 
       %{report: report}
     end
@@ -759,7 +768,8 @@ defmodule Lightning.UsageTrackingTest do
 
       assert %{
                submitted: true,
-               submitted_at: submitted_at
+               submitted_at: submitted_at,
+               submission_status: :success
              } = Repo.get!(Report, report.id)
 
       assert DateTime.diff(DateTime.utc_now(), submitted_at, :second) < 2
@@ -772,7 +782,8 @@ defmodule Lightning.UsageTrackingTest do
 
       assert %{
                submitted: true,
-               submitted_at: submitted_at
+               submitted_at: submitted_at,
+               submission_status: :success
              } = updated_report
 
       assert DateTime.diff(DateTime.utc_now(), submitted_at, :second) < 2
@@ -781,14 +792,12 @@ defmodule Lightning.UsageTrackingTest do
 
   describe ".update_report_submission - unsuccessful" do
     setup do
-      # Create an `artificial` report to validate logic - there is currently
-      # no use case that would result in code overwriting a report that has
-      # been successfully submitted,
       report =
         insert(
           :usage_tracking_report,
-          submitted: true,
-          submitted_at: DateTime.utc_now()
+          submitted: false,
+          submitted_at: DateTime.utc_now(),
+          submission_status: :pending
         )
 
       %{report: report}
@@ -801,7 +810,8 @@ defmodule Lightning.UsageTrackingTest do
 
       assert %{
                submitted: false,
-               submitted_at: nil
+               submitted_at: nil,
+               submission_status: :failure
              } = Repo.get!(Report, report.id)
     end
 
@@ -812,7 +822,8 @@ defmodule Lightning.UsageTrackingTest do
 
       assert %{
                submitted: false,
-               submitted_at: nil
+               submitted_at: nil,
+               submission_status: :failure
              } = updated_report
     end
   end